+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - unreleased
+
+### Added
+- Added an option to the optuna search to use multiprocessing using the suggestions made in 
+  https://github.com/optuna/optuna/issues/2862 .
+  This has not been extensively tested in real projects.
+  Therefore, use with care and please report any issues you encounter.
+
 ## [0.12.2] - 2022-12-14
 
 ### Fixed